@@ -1,5 +1,6 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { Subscription, forkJoin } from 'rxjs';
+import { Subscription, forkJoin, Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
 import { MatButtonModule } from "@angular/material/button";
 import { DemoFlipperService } from '../../services/demo-flipper.service';
 import { FormsModule } from '@angular/forms';
@@ -12,25 +13,26 @@ import { MatSelectModule } from '@angular/material/select';
 import { MatIconModule } from '@angular/material/icon';
 import { BlanketBlasterModule } from '../blanket-blaster.module';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { DateTimeService } from '../../services/date-time.service';
+import { PollingService } from '../../services/polling-service.service';
 
 @Component({
-  selector: 'ct-six-lane-flipper',
-  templateUrl: './six-lane-flipper.component.html',
-  styleUrls: ['./six-lane-flipper.component.scss'],
-  standalone: true,
-  imports: [
-    CommonModule,
-    FormsModule,
-    MatButtonModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatDatepickerModule,
-    MatNativeDateModule,
-    MatSelectModule,
-    MatIconModule,
-    BlanketBlasterModule,
-    MatSlideToggleModule
-  ]
+    selector: 'ct-six-lane-flipper',
+    templateUrl: './six-lane-flipper.component.html',
+    styleUrls: ['./six-lane-flipper.component.scss'],
+    imports: [
+        CommonModule,
+        FormsModule,
+        MatButtonModule,
+        MatFormFieldModule,
+        MatInputModule,
+        MatDatepickerModule,
+        MatNativeDateModule,
+        MatSelectModule,
+        MatIconModule,
+        BlanketBlasterModule,
+        MatSlideToggleModule
+    ]
 })
 export class SixLaneFlipperComponent implements OnInit, OnDestroy {
   sub: Subscription;
@@ -40,7 +42,6 @@ export class SixLaneFlipperComponent implements OnInit, OnDestroy {
   subDaily: Subscription;
   subStatus: Subscription;
 
-  selectedDate: string | null = null;
   serialNumbers = [67808, 67806, 67807, 67805, 67804, 67803];
 
   public liveMode: boolean = false;
@@ -48,9 +49,15 @@ export class SixLaneFlipperComponent implements OnInit, OnDestroy {
   public realDataArray: any[] = [];
 
   private pollingTimeout: any = null;
+  private destroy$ = new Subject<void>();
+  private readonly POLLING_INTERVAL = 6000; // 6 seconds
 
-  constructor(private demoFlipperService: DemoFlipperService) {}
-
+  constructor(
+    private demoFlipperService: DemoFlipperService,
+    private dateTimeService: DateTimeService,
+    private pollingService: PollingService
+  ) {}
+  
   ident(index: number, lane: any): number {
     return index;
   }
@@ -61,8 +68,43 @@ export class SixLaneFlipperComponent implements OnInit, OnDestroy {
     this.subHourly = this.createDummySub(1000);
     this.subDaily = this.createDummySub(1000);
     this.subStatus = this.createDummySub(500);
-  }
 
+    // Initial setup with DateTimeService
+    const isLive = this.dateTimeService.getLiveMode();
+    const wasConfirmed = this.dateTimeService.getConfirmed();
+
+    if (!isLive && wasConfirmed) {
+      this.fetchData();
+    }
+  
+    // Subscribe to live mode changes
+    this.dateTimeService.liveMode$
+      .pipe(takeUntil(this.destroy$))
+      .subscribe((isLive: boolean) => {
+        this.liveMode = isLive;
+  
+        if (this.liveMode) {
+          this.fetchData();
+          this.startLivePolling();
+        } else {
+          this.stopLivePolling();
+          this.realDataArray = [];
+        }
+      });
+  
+    // Subscribe to manual confirm from modal
+    this.dateTimeService.confirmTrigger$
+      .pipe(takeUntil(this.destroy$))
+      .subscribe(() => {
+        this.liveMode = false;
+        this.stopLivePolling();
+        this.fetchData();
+      });
+
+    // Initial data fetch
+    this.fetchData();
+  }
+  
   ngOnDestroy() {
     if (this.sub) this.sub.unsubscribe();
     if (this.subFiveMinutes) this.subFiveMinutes.unsubscribe();
@@ -71,55 +113,23 @@ export class SixLaneFlipperComponent implements OnInit, OnDestroy {
     if (this.subDaily) this.subDaily.unsubscribe();
     if (this.subStatus) this.subStatus.unsubscribe();
     this.stopLivePolling();
+    this.destroy$.next();
+    this.destroy$.complete();
   }
 
   private createDummySub(interval: number): Subscription {
-    return new Subscription(); // Replace with logic if needed
+    return new Subscription();
   }
 
   fetchData() {
-    const date = this.selectedDate || new Date().toISOString().split('T')[0];
+    const date = this.dateTimeService.getStartTime()?.split('T')[0] || new Date().toISOString().split('T')[0];
     const apiCalls = this.serialNumbers.map(serial =>
       this.demoFlipperService.getLiveEfficiencySummary(serial, date)
     );
 
-    forkJoin(apiCalls).subscribe({
-      next: (results: any[]) => {
-        this.realDataArray = results.flatMap((result, i) => {
-          const serial = this.serialNumbers[i];
-          if (!result || !Array.isArray(result.flipperData)) {
-            console.warn(`Invalid flipperData for serial ${serial}:`, result);
-            return [];
-          }
-          return result.flipperData;
-        });
-        this.hasFetchedData = true;
-      },
-      error: (err) => {
-        console.error("Error during fetchData():", err);
-      }
-    });
-  }
-
-  onLiveModeChange(checked: boolean) {
-    this.liveMode = checked;
-
-    if (this.liveMode) {
-      this.startLivePolling();
-    } else {
-      this.stopLivePolling();
-    }
-  }
-
-  private startLivePolling() {
-    const dateToUse = this.selectedDate || new Date().toISOString().split('T')[0];
-
-    const poll = () => {
-      const apiCalls = this.serialNumbers.map(serial =>
-        this.demoFlipperService.getLiveEfficiencySummary(serial, dateToUse)
-      );
-
-      forkJoin(apiCalls).subscribe({
+    forkJoin(apiCalls)
+      .pipe(takeUntil(this.destroy$))
+      .subscribe({
         next: (results: any[]) => {
           this.realDataArray = results.flatMap((result, i) => {
             const serial = this.serialNumbers[i];
@@ -130,22 +140,50 @@ export class SixLaneFlipperComponent implements OnInit, OnDestroy {
             return result.flipperData;
           });
           this.hasFetchedData = true;
-
-          if (this.liveMode) {
-            this.pollingTimeout = setTimeout(poll, 1000); // 🔁 wait 1 second after each call
-          }
         },
         error: (err) => {
-          console.error("Polling error:", err);
-
-          if (this.liveMode) {
-            this.pollingTimeout = setTimeout(poll, 1000); // 🔁 retry after 1s
-          }
+          console.error("Error during fetchData():", err);
         }
       });
+  }
+
+  private startLivePolling() {
+    const date = this.dateTimeService.getStartTime()?.split('T')[0] || new Date().toISOString().split('T')[0];
+
+    const poll = () => {
+      const apiCalls = this.serialNumbers.map(serial =>
+        this.demoFlipperService.getLiveEfficiencySummary(serial, date)
+      );
+
+      forkJoin(apiCalls)
+        .pipe(takeUntil(this.destroy$))
+        .subscribe({
+          next: (results: any[]) => {
+            this.realDataArray = results.flatMap((result, i) => {
+              const serial = this.serialNumbers[i];
+              if (!result || !Array.isArray(result.flipperData)) {
+                console.warn(`Invalid flipperData for serial ${serial}:`, result);
+                return [];
+              }
+              return result.flipperData;
+            });
+            this.hasFetchedData = true;
+
+            if (this.liveMode) {
+              this.pollingTimeout = setTimeout(poll, this.POLLING_INTERVAL);
+            }
+          },
+          error: (err) => {
+            console.error("Polling error:", err);
+
+            if (this.liveMode) {
+              this.pollingTimeout = setTimeout(poll, this.POLLING_INTERVAL);
+            }
+          }
+        });
     };
 
-    poll(); // start initial poll
+    poll();
   }
 
   private stopLivePolling() {
