@@ -1,4 +1,4 @@
-import { Component, inject, Output, Input, EventEmitter } from '@angular/core';
+import { Component, inject, Output, Input, EventEmitter, ViewChild } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Router } from '@angular/router';
 
@@ -11,6 +11,7 @@ import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatListModule } from '@angular/material/list';
 import { MatIconModule } from '@angular/material/icon';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { MatMenuModule, MatMenuTrigger } from '@angular/material/menu';
 
 import { trigger, state, style, animate, transition, query, group } from '@angular/animations';
 
@@ -18,6 +19,7 @@ import { Observable } from 'rxjs';
 import { map, shareReplay } from 'rxjs/operators';
 
 import { UserService } from '../user.service';
+import { DateTimeModalComponent } from '../components/date-time-modal/date-time-modal.component';
 
 const left = [
   query(':enter, :leave', style({ position: 'absolute', width: '15em' })),
@@ -36,35 +38,38 @@ const right = [
 ];
 
 @Component({
-  selector: 'nav-main-menu',
-  templateUrl: './nav-main-menu.component.html',
-  styleUrl: './nav-main-menu.component.scss',
-  standalone: true,
-  imports: [
-    CommonModule,
-    RouterOutlet,
-    RouterModule,
-    MatToolbarModule,
-    MatButtonModule,
-    MatSidenavModule,
-    MatListModule,
-    MatIconModule,
-    MatSlideToggleModule
-  ],
-  animations: [
-    trigger('menuSlider', [
-      transition(':increment', right),
-      transition(':decrement', left),
-    ]),
-    trigger('menuSwap', [
-      transition(':enter', [style({ transform: 'translateX(-15em)', position: 'absolute' }), animate('.3s ease-out', style({ transform: 'translateX(0%)', position: 'absolute' }))]),
-      transition(':leave', [style({ transform: 'translateX(0%)', position: 'absolute' }), animate('.3s ease-out', style({ transform: 'translateX(-15em)', position: 'absolute' }))])
-    ])
-  ]
+    selector: 'nav-main-menu',
+    templateUrl: './nav-main-menu.component.html',
+    styleUrl: './nav-main-menu.component.scss',
+    imports: [
+        CommonModule,
+        RouterOutlet,
+        RouterModule,
+        MatToolbarModule,
+        MatButtonModule,
+        MatSidenavModule,
+        MatListModule,
+        MatIconModule,
+        MatSlideToggleModule,
+        MatMenuModule,
+        DateTimeModalComponent
+    ],
+    animations: [
+        trigger('menuSlider', [
+            transition(':increment', right),
+            transition(':decrement', left),
+        ]),
+        trigger('menuSwap', [
+            transition(':enter', [style({ transform: 'translateX(-15em)', position: 'absolute' }), animate('.3s ease-out', style({ transform: 'translateX(0%)', position: 'absolute' }))]),
+            transition(':leave', [style({ transform: 'translateX(0%)', position: 'absolute' }), animate('.3s ease-out', style({ transform: 'translateX(-15em)', position: 'absolute' }))])
+        ])
+    ]
 })
 export class NavMainMenuComponent {
   @Output() darkModeToggleEvent = new EventEmitter();
   @Input() isDarkMode: boolean;
+  @ViewChild('dateMenuTrigger') dateMenu: MatMenuTrigger;
+
 
   private breakpointObserver = inject(BreakpointObserver);
 
@@ -133,4 +138,15 @@ export class NavMainMenuComponent {
     this.shownMenu = this.menuHistory.pop() || '';
     this.menuIndex--;
   }
+
+  onDateTimeModalClose(): void {
+    setTimeout(() => {
+      if (this.dateMenu) {
+        this.dateMenu.closeMenu();
+        console.log('confirm button clicked');
+      }
+    });
+  }
+  
+  
 }
