@@ -1,17 +1,16 @@
 import { Component } from "@angular/core";
 import { CommonModule } from "@angular/common";
-import { SixLaneFlipperComponent } from "../blanket-blaster/six-lane-flipper/six-lane-flipper.component";
+import { BlanketBlasterModule } from "../blanket-blaster/blanket-blaster.module";
 
 @Component({
-  selector: "app-test",
-  standalone: true,
-  imports: [
-    CommonModule,
-    SixLaneFlipperComponent
-  ],
-  templateUrl: "./test.component.html",
-  styleUrls: ["./test.component.scss"],
+    selector: "app-test",
+    imports: [
+        CommonModule,
+        BlanketBlasterModule
+    ],
+    templateUrl: "./test.component.html",
+    styleUrls: ["./test.component.scss"]
 })
 export class TestComponent {
-  // Component is now empty as it just serves as a container for the six-lane-flipper
+  // Component is now empty as it just serves as a container for the demo-flipper
 }
